Refactor logic with lexers and Option
Each of the token lexers should return `Token`, since there's no
possibility of failure if it works. The parser itself failing is
the failure mode. Instead, I add a wrapper `lexValid` that
provides an alternative for each of the valid lexers, and then
the failure of that results in calling `lexNoOp`.